rpmsg: smd: Use spinlock in tx path
Backporting the change from 4.19 kernel to 4.14 kernel.

By switching the tx_lock to a spinlock we allow clients to use
rpmsg_trysend() from atomic context.

The mutex was interruptable as it was previously held for the duration
of some client waiting for available space in the FIFO, but this was
recently changed to only be held temporarily - allowing us to replace it
with a spinlock.

Change-Id: Ib645758b7c5f7e8b475dfa0c7dab4b11d30ea6b9
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Git-commit: 33e3820dda8876792bd8135db633c741a07263be
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Prudhvi Yarlagadda <pyarlaga@codeaurora.org>