qcacld-3.0: Remove DPU signatures from TDLS Add STA Response
The ucastSig and bcastSig fields in tSirTdlsAddStaRsp date back
to a previous version of the driver that used these signatures to
synchronize the data plane with the control plane. However the current
version of the driver does not use these parameters, so remove them as
the final stage of the plan to completely remove these signatures from
throughout the driver.

Since the legacy tSirTdlsAddStaRsp data structure must exactly match
the converged struct tdls_add_sta_rsp, this change is co-dependent
with the qca-wifi-host-cmn change:
  qcacmn: Remove DPU signatures from TDLS Add STA Response
  Change-Id: I7750fac9e4870d75358606bea427a5dbab97c665

Change-Id: I765b090fc5dbe9bdaf83b05a72223c0a12f71a18
CRs-Fixed: 2200996