qcacld-3.0: Fix bad typecast in pe_mc_process_handler()
In pe_mc_process_handler() the mac_ctx is currently typecast to a
tHalHandle when calling pe_process_messages(). However
pe_process_messages() actually expects a tpAniSirGlobal, and the
typecast results in a build failure when strict type checking is
enabled. To fix the build failure remove the typecast.

Change-Id: Ie8a38845f0e2bf76205326a1b5fe7691a8f8de12
CRs-Fixed: 2254952