qcacld-3.0: Fix bad csr_parse_bss_description_ies() typecast
In csr_roam_call_callback() there is a call to
csr_parse_bss_description_ies() where the MAC context is typecast to
tHalHandle as part of the function call. However this typecast is
incorrect since csr_parse_bss_description_ies() expects to get a MAC
context. Due to the manner in which tHalHandle is defined this does
not current cause a build issue, but an upcoming change will enforce
type checking so update this call to remove the inappropriate
typecasts.

Change-Id: I16e1f33e6f00ed1c610511a60f5974f0f0d71cd5
CRs-Fixed: 2268807