qcacld-3.0: Remove is_vdev_valid flag from the vdev structure wma_txrx_node
The flag is_vdev_valid and vdev_active serve the same purpose,
they are used to prevent the race condition where wmi vdev level
command gets posted to FW after vdev is deleted. Both the flags
are set to true in wma_vdev_attach when the vdev is created and
set to false in wma_vdev_detach where the vdev delete command is
sent to fw.

So remove is_vdev_valid flag and replace it with the existing
vdev_active flag.

Change-Id: I5199bc7d473470fd9f16320d323f86257a24ae19
CRs Fixed: 2312321