qcacld-3.0: Limit cherry-picked change Ids to 50
The change Ids of the most recent cherry-picked git commits are appended
to the build tag. In some cases, we have seen build tags containing
thousands of change Ids. To avoid this, limit the number of change Ids
to 50. This should be sufficient to know if some recent change is in the
build or not, without including an unbounded number of change Ids in the
build tag.

Change-Id: I3713cb8e095730ced44d88f7d05a92c2dc77f8b0
CRs-Fixed: 2311431